Deprecate unnecessary Helm Chart deployment (#4507)
