+++ tests/e2e/tidbcluster/stability-asts.go
@@ -119,7 +119,7 @@ var _ = ginkgo.Describe("[Stability]", func() {
 				WebhookEnabled:    true,
 				StsWebhookEnabled: false,
 			}
-			oa = tests.NewOperatorActions(cli, c, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, ocfg, e2econfig.TestConfig, nil, fw, f)
+			oa = tests.NewOperatorActions(cli, c, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, ocfg, e2econfig.TestConfig, fw, f)
 			ginkgo.By("Installing CRDs")
 			oa.CleanCRDOrDie()
 			oa.CreateCRDOrDie(ocfg)
@@ -327,7 +327,7 @@ var _ = ginkgo.Describe("[Stability]", func() {
 			ImagePullPolicy: v1.PullIfNotPresent,
 			TestMode:        true,
 		}
-		oa = tests.NewOperatorActions(cli, c, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, ocfg, e2econfig.TestConfig, nil, fw, f)
+		oa = tests.NewOperatorActions(cli, c, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, ocfg, e2econfig.TestConfig, fw, f)
 		ginkgo.By("Installing CRDs")
 		oa.CleanCRDOrDie()
 		oa.CreateCRDOrDie(ocfg)
@@ -419,7 +419,7 @@ var _ = ginkgo.Describe("[Stability]", func() {
 			ImagePullPolicy: v1.PullIfNotPresent,
 			TestMode:        true,
 		}
-		oa = tests.NewOperatorActions(cli, c, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, ocfg, e2econfig.TestConfig, nil, fw, f)
+		oa = tests.NewOperatorActions(cli, c, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, ocfg, e2econfig.TestConfig, fw, f)
 		ginkgo.By("Installing CRDs")
 		oa.CleanCRDOrDie()
 		oa.ReplaceCRDOrDie(ocfg)
